LayoutTests/fast/viewport/ios/wide-desktop-viewport-in-enhanced-windowing-mode.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true enhancedWindowingEnabled=true ShouldIgnoreMetaViewport=true ] -->
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<style>
body, html {
    margin: 0;
    width: 100%;
    height: 100%;
}

.horizontal-bar {
    background: linear-gradient(to right, red 0%, green 50%, blue 100%);
    width: 1280px;
    height: 100px;
}
</style>
<script>
jsTestIsAsync = true;
addEventListener("load", async () => {
    scale = visualViewport.scale;
    contentWidth = document.scrollingElement.scrollWidth;
    // The expected value of 820 here matches the minimum shrink-to-fit width when preferring
    // horizontal scrolling.
    shouldBe("contentWidth", "1280");
    shouldBe("contentWidth * scale", "820");
    finishJSTest();
});
</script>
</head>
<body>
<div class="horizontal-bar"></div>
</body>
</html>
