LayoutTests/fast/viewport/ios/wide-desktop-viewport-in-enhanced-windowing-mode-expected.txt

PASS contentWidth is 1280
PASS contentWidth * scale is 820
PASS successfullyParsed is true

TEST COMPLETE
